Moving methods of SQL backends
