+++ ibis/backends/postgres/__init__.py
@@ -1,12 +1,17 @@
 """PostgreSQL backend."""
+import contextlib
+
 from ibis.backends.base.sql.alchemy import BaseAlchemyBackend
 
 from .client import PostgreSQLClient
+from .compiler import PostgreSQLCompiler
+from . import udf
 
 
 class Backend(BaseAlchemyBackend):
     name = 'postgres'
     client_class = PostgreSQLClient
+    compiler = PostgreSQLCompiler
 
     def connect(
         self,
@@ -108,3 +113,76 @@ class Backend(BaseAlchemyBackend):
         if database == self.current_database:
             database = None
         return super().list_tables(like, database)
+
+    @contextlib.contextmanager
+    def begin(self):
+        with super().begin() as bind:
+            previous_timezone = bind.execute('SHOW TIMEZONE').scalar()
+            bind.execute('SET TIMEZONE = UTC')
+            try:
+                yield bind
+            finally:
+                bind.execute(f"SET TIMEZONE = '{previous_timezone}'")
+
+    def table(self, name, database=None, schema=None):
+        """Create a table expression that references a particular a table
+        called `name` in a PostgreSQL database called `database`.
+
+        Parameters
+        ----------
+        name : str
+            The name of the table to retrieve.
+        database : str, optional
+            The database in which the table referred to by `name` resides. If
+            ``None`` then the ``current_database`` is used.
+        schema : str, optional
+            The schema in which the table resides.  If ``None`` then the
+            `public` schema is assumed.
+
+        Returns
+        -------
+        table : TableExpr
+            A table expression.
+        """
+        if database is not None and database != self.current_database:
+            return self.database(name=database).table(name=name, schema=schema)
+        else:
+            alch_table = self._get_sqla_table(name, schema=schema)
+            node = self.table_class(alch_table, self, self._schemas.get(name))
+            return self.table_expr_class(node)
+
+    def udf(
+        self, pyfunc, in_types, out_type, schema=None, replace=False, name=None
+    ):
+        """Decorator that defines a PL/Python UDF in-database based on the
+        wrapped function and turns it into an ibis function expression.
+
+        Parameters
+        ----------
+        pyfunc : function
+        in_types : List[ibis.expr.datatypes.DataType]
+        out_type : ibis.expr.datatypes.DataType
+        schema : str
+            optionally specify the schema in which to define the UDF
+        replace : bool
+            replace UDF in database if already exists
+        name: str
+            name for the UDF to be defined in database
+
+        Returns
+        -------
+        Callable
+
+        Function that takes in ColumnExpr arguments and returns an instance
+        inheriting from PostgresUDFNode
+        """
+
+        return udf(
+            client=self,
+            python_func=pyfunc,
+            in_types=in_types,
+            out_type=out_type,
+            schema=schema,
+            replace=replace,
+            name=name,
+        )
