+++ ibis/backends/tests/base.py
@@ -150,4 +150,4 @@ class BackendTest(abc.ABC):
     def make_context(
         self, params: Optional[Mapping[ir.ValueExpr, Any]] = None
     ):
-        return self.api.client_class.compiler.make_context(params=params)
+        return self.api.compiler.make_context(params=params)
