+++ ibis/tests/expr/mocks.py
@@ -18,7 +18,7 @@ from typing import Optional
 import pytest
 
 import ibis.expr.types as ir
-from ibis.backends.base.sql import SQLClient
+from ibis.backends.base.sql import SQLClient, BaseSQLBackend
 from ibis.backends.base.sql.alchemy import (
     AlchemyCompiler,
     AlchemyTable,
@@ -28,8 +28,31 @@ from ibis.expr.schema import Schema
 from ibis.expr.typing import TimeContext
 
 
+class MockBackend(BaseSQLBackend):
+    name = 'mock'
+    version = '1.0'
+    client_class = None
+    current_database = 'mockdb'
+
+    def __init__(self, client):
+        self.client = client
+
+    def connect(self):
+        pass
+
+    def list_tables(self):
+        return list(self.client._tables)
+
+    def list_databases(self):
+        return ['mockdb']
+
+    def fetch_from_cursor(self):
+        pass
+
+
 class MockConnection(SQLClient, metaclass=abc.ABCMeta):
     def __init__(self):
+        self.backend = MockBackend(client=self)
         self.executed_queries = []
 
     _tables = {
